+++ lib/src/generator/generator_frontend.dart
@@ -27,6 +27,13 @@ class GeneratorFrontEnd implements Generator {
             'longer be supported.',
       );
     }
+    if (_generatorBackend.options.resourcesDir != null) {
+      packageGraph?.defaultPackage.warn(
+        PackageWarning.deprecated,
+        message: "The '--resources-dir' option is deprecated, and will soon be "
+            'removed.',
+      );
+    }
 
     var indexElements = packageGraph == null
         ? const <Indexable>[]